Some general backend cleanup. moving packages around, removing some traces of electron, and prepping for standardised object store library.
