+++ packages/auth/package.json
@@ -6,15 +6,20 @@
   "author": "Budibase",
   "license": "AGPL-3.0",
   "dependencies": {
+    "aws-sdk": "^2.901.0",
     "bcryptjs": "^2.4.3",
     "ioredis": "^4.27.1",
     "jsonwebtoken": "^8.5.1",
     "koa-passport": "^4.1.4",
+    "node-fetch": "^2.6.1",
     "passport-google-auth": "^1.0.2",
     "passport-google-oauth": "^2.0.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "uuid": "^8.3.2"
+    "sanitize-s3-objectkey": "^0.0.1",
+    "tar-fs": "^2.1.1",
+    "uuid": "^8.3.2",
+    "zlib": "^1.0.5"
   },
   "devDependencies": {
     "ioredis-mock": "^5.5.5"
