+++ packages/server/package.json
@@ -90,16 +90,12 @@
     "arangojs": "7.2.0",
     "aws-sdk": "^2.767.0",
     "bcryptjs": "2.4.3",
-    "bull": "^3.22.0",
-    "bull-board": "^1.5.1",
+    "bull": "^3.22.4",
+    "bull-board": "^2.0.1",
     "chmodr": "1.2.0",
     "csvtojson": "2.0.10",
     "dotenv": "8.2.0",
     "download": "8.0.0",
-    "electron-is-dev": "1.2.0",
-    "electron-unhandled": "3.0.2",
-    "electron-updater": "4.3.1",
-    "electron-util": "0.14.2",
     "fix-path": "3.0.0",
     "fs-extra": "8.1.0",
     "jimp": "0.16.1",
@@ -125,10 +121,8 @@
     "pouchdb-all-dbs": "1.0.2",
     "pouchdb-find": "^7.2.2",
     "pouchdb-replication-stream": "1.2.9",
-    "sanitize-s3-objectkey": "0.0.1",
     "server-destroy": "1.0.1",
     "svelte": "3.30.0",
-    "tar-fs": "2.1.0",
     "to-json-schema": "0.2.5",
     "uuid": "3.3.2",
     "validate.js": "0.13.1",
@@ -140,10 +134,8 @@
     "@budibase/standard-components": "^0.8.16",
     "@jest/test-sequencer": "^24.8.0",
     "docker-compose": "^0.23.6",
-    "electron": "10.1.3",
-    "electron-builder": "^22.9.1",
-    "electron-builder-notarize": "^1.1.2",
     "eslint": "^6.8.0",
+    "express": "^4.17.1",
     "jest": "^24.8.0",
     "nodemon": "^2.0.4",
     "pouchdb-adapter-memory": "^7.2.1",
