+++ packages/server/src/constants/index.js
@@ -1,5 +1,6 @@
 const { BUILTIN_ROLE_IDS } = require("../utilities/security/roles")
 const { UserStatus } = require("@budibase/auth").constants
+const { ObjectStoreBuckets } = require("@budibase/auth").objectStore
 
 exports.LOGO_URL =
   "https://d33wubrfki0l68.cloudfront.net/aac32159d7207b5085e74a7ef67afbb7027786c5/2b1fd/img/logo/bb-emblem.svg"
@@ -89,8 +90,5 @@ exports.BaseQueryVerbs = {
   DELETE: "delete",
 }
 
-exports.ObjectStoreBuckets = {
-  BACKUPS: "backups",
-  APPS: "prod-budi-app-assets",
-  TEMPLATES: "templates",
-}
+// pass through the list from the auth/core lib
+exports.ObjectStoreBuckets = ObjectStoreBuckets
