+++ packages/server/src/utilities/budibaseDir.js
@@ -1,11 +1,10 @@
 const { join } = require("./centralPath")
-const { homedir, tmpdir } = require("os")
+const { homedir } = require("os")
 const env = require("../environment")
+const { budibaseTempDir } = require("@budibase/auth").objectStore
 
 module.exports.budibaseAppsDir = function () {
   return env.BUDIBASE_DIR || join(homedir(), ".budibase")
 }
 
-module.exports.budibaseTempDir = function () {
-  return join(tmpdir(), ".budibase")
-}
+module.exports.budibaseTempDir = budibaseTempDir
